format api call into user_route rank search
